+++ API_Robot/motor.h
@@ -9,7 +9,7 @@ class Motor
 {
 public:
 	Motor(int motor);
-	void setVelocity(float velocity);
+	void setVelocity(double velocity);
 	//setVoltage();
 	//void enableMotors();
 
@@ -17,10 +17,10 @@ private:
 	int _motor;
 	int _dir_pin;
 	int _pwm_pin;
-	float _velocity;
+	double _velocity;
 	int _pwm;
 };
 
-int vel2pwm(float velocity);
+int vel2pwm(double velocity);
 
 #endif
